requirements.txt
# 飞书官方SDK
lark-oapi>=1.5.3

# 工具库
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.0.0
psutil>=5.9.0

# 注意:
# Claude Code CLI 需要单独安装: npm install -g @anthropic-ai/claude-code
